LayoutTests/fast/mediastream/mediastreamtrack-configurationchange-while-muted-expected.txt



PASS configurationchange is deferred while muted and fires after unmute
PASS multiple configurationchange notifications while muted coalesce into one event on unmute
